feat: many things

+++ src/Entity/Product.php
@@ -27,7 +27,7 @@ class Product
 
     #[ORM\Column]
     #[Groups(['product_simple'])]
-    private ?int $price = null;
+    private ?float $price = null;
 
     /**
      * @var Collection<int, subCategory>
@@ -90,12 +90,12 @@ class Product
         return $this;
     }
 
-    public function getPrice(): ?int
+    public function getPrice(): ?float
     {
         return $this->price;
     }
 
-    public function setPrice(int $price): static
+    public function setPrice(float $price): static
     {
         $this->price = $price;
